rewrite in TypeScript.
