+++ .gitignore
@@ -7,3 +7,4 @@ run/
 .DS_Store
 *.swp
 yarn.lock
+dist/
