+++ .travis.yml
@@ -1,10 +1,12 @@
 sudo: false
 language: node_js
 node_js:
-  - '4'
-  - '6'
+  - node
+before_script:
+  - npm install -g gulp
 install:
   - npm i npminstall && npminstall
+  - gulp
 script:
   - npm run ci
 after_script:
