--- lib/api/media.js
@@ -1,66 +1,0 @@
-'use strict';
-
-const assert = require('assert');
-const fs = require('fs');
-const path = require('path');
-
-/**
- * 媒体相关 API
- * @see https://open-doc.dingtalk.com/doc2/detail.htm?articleId=104971&docType=1
- * @type {Media}
- */
-module.exports = class Media {
-  constructor(client, options) {
-    this.client = client;
-    this.options = options;
-  }
-
-  /**
-   * 上传媒体文件
-   *  - media/upload
-   * @param {String} type 媒体文件类型, image/voice/file
-   * @param {String} filePath 文件路径
-   * @return {Object} 操作结果, { media_id, type, created_at }
-   */
-  * upload(type, filePath) {
-    assert(type, 'type required');
-    assert(filePath, 'filePath required');
-
-    const accessToken = yield this.client.getAccessToken();
-    const url = `${this.options.host}/media/upload?access_token=${accessToken}&type=${type}`;
-    return yield this.client.upload(url, { field: 'media', path: filePath });
-  }
-
-  /**
-   * 获取媒体文件
-   * 注意: 资源不存在时, 钉钉返回的是 200 和 空内容, 而不是文档描述的 json
-   * 图片缩放参见: https://open-doc.dingtalk.com/doc2/detail.htm?treeId=176&articleId=104948&docType=1
-   * @param {String} id - 资源ID
-   * @return {String} 资源下载地址
-   */
-  * get(id) {
-    const response = yield this.client.get('media/get', { media_id: id });
-    if (response.status === 301 || response.status === 302) {
-      return response.headers.location;
-    }
-  }
-
-  /**
-   * 下载媒体文件
-   * @param {String} id - 资源ID
-   * @param {String} targetDir - 下载的目录, 目录必须存在.
-   * @param {String} [fileName] - 文件名, 为空时直接使用服务器上的文件名
-   * @return {String} 本地文件地址
-   */
-  * download(id, targetDir, fileName) {
-    const url = yield this.get(id);
-    const filePath = path.join(targetDir, fileName ? fileName : path.basename(url));
-    if (url) {
-      yield this.client.urllib.request(url, {
-        writeStream: fs.createWriteStream(filePath),
-        timeout: 20000,
-      });
-      return filePath;
-    }
-  }
-};
